txtorcon endpoints demo
This demonstrates using txtorcon (v0.10.0 or greater) along with a Twisted endpoint-enabled application (in this case, the built-in Web server) to launch a freshly-created hidden service (in this case, a Web site).

You could use similar techniques to make any endpoint-aware Twisted application available as a hidden service. If you need more control over Tor's configuration, you can write a few lines of Python to achieve the same effect (but with a custom Tor configuration).

https://www.github.com/meejah/txtorcon ( https://www.github.com/meejah/txtorcon )