Building a password generator with python.
In this video we build a password generator in python3.x. Usually these things are written using compiled languages for more speed, but this can be used as a guideline regardless. This should give you a great understanding of how python syntax looks, and how password generators work. There are many directions we can go with this from here.

Use the code provided here freely!

** I appologize for not showing the entire code at the end. At the start of minute 34 is where I put the final touches on the code.